Python Script
In this task, you will create simple GUI interfaces with Raspberry Pi. RPi supports python,
which has many GUI packages including Qt, TkInter and Kivy. As before, while this task is
focusing on Python, you can use any programming language you are most comfortable with
(and supported by RPi platform) to perform this task and future tasks incorporating a RPi. For
example, if you are using C++, Qt has a complete GUI package for C++ which is very similar
to the Python package. In this task, based on what you learnt from the video tutorials, you are
required to build a simple GUI interface to turn on one of three LED lights.